GOCI-2335 variant_doc_fix
