+++ scripts/generate_solr_docs.py
@@ -15,7 +15,7 @@ from ols import OLSData
 from document_types import publication
 from document_types import trait
 from document_types import study
-from document_types import Variant
+from document_types import variant
 
 
 def publication_data(connection, limit=0):
@@ -70,74 +70,8 @@ def save_data(data, data_type=None):
     with open(path, 'w') as outfile:
         outfile.write(jsonData)
 
-
-
-def get_variant_data(DATABASE_NAME, limit=0):
-    '''
-    Get Variant data for Solr document.
-    '''
-
-    # function to retrieve further data from the database:
-    def get_more_variant_data(row):
-        # Extracting basic variant information:
-        resourcename = 'variant'
-        ID = row['ID']
-        rsID = row['RS_ID']
-        consequence = row['FUNCTIONAL_CLASS']
-
-        # Extracting genomic location:
-        location = variant_cls.get_variant_location(ID)
-
-        # Extracting mapped genes:
-        mapped_genes_list = variant_cls.get_mapped_genes(ID)
-        
-        mapped_genes_names = [x.split("|")[0] for x in mapped_genes_list]
-
-        # Extracting association count:
-        associations = variant_cls.get_association_count(ID)
-        
-        # Combining data into a dictionary:
-        varDoc = {
-            'resourcename' : resourcename,
-            'id' : "%s-%s" % (resourcename,ID),
-            'title' : rsID,
-            'rsID' : rsID,
-            'associationCount' : associations,
-            'mappedGenes' : mapped_genes_list,
-            'chromosomeName' : location['chromosome'],
-            'chromosomePosition' : location['position'],
-            'region' : location['region'],
-            'consequence' : consequence,
-            'link' : 'variants/rs7329174'
-
-        }
-        varDoc['description'] =  '%s:%s, %s, %s, mapped to: %s, associaitons: %s' %(
-                    varDoc['chromosomeName'], varDoc['chromosomePosition'], varDoc['region'], varDoc['consequence'],
-                    ",".join(mapped_genes_names),varDoc['associationCount']
-                )
-
-        # Adding to document list:
-        all_variant_data.append(varDoc)
-
-    # Initialize empty list for the documents:
-    all_variant_data = []
-
-    # Step 1: initialize variant object:
-    variant_cls = Variant.Variant(DATABASE_NAME)
-
-    # Step 2: retrieve all the variants in the database:
-    variants_df = variant_cls.get_snps()
-
-    # Inintialize progress bar:
-    tqdm.pandas(desc="Returning variant data")
-
-    # Step 3: Calling apply to retrieve all variant data:
-    if limit != 0:
-        variants_df[1:limit].progress_apply(get_more_variant_data, axis = 1)
-    else:
-        variants_df.progress_apply(get_more_variant_data, axis = 1)
-
-    return all_variant_data
+def variant_data(connection, limit=0):
+    return variant.get_variant_data(connection, limit)
 
 
 def get_gene_data():
@@ -318,10 +252,6 @@ def get_gene_data():
         print exception
 
 
-
-
-
-
 if __name__ == '__main__':
     '''
     Create Solr documents for categories of interest.
@@ -331,7 +261,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--database', default='SPOTREL', choices=['DEV3', 'SPOTREL'],
                         help='Run as (default: SPOTREL).')
-    parser.add_argument('--limit', type=int, help='Limit the nuber of document to this number for testing purposes.')
+    parser.add_argument('--limit', type=int, help='Limit the number of created documents to this number for testing purposes.')
     parser.add_argument('--document', default='publication',
                         choices=['publication', 'trait', 'variant', 'all'],
                         help='Run as (default: publication).')
@@ -354,7 +284,7 @@ if __name__ == '__main__':
         'publication' : publication_data, 
         # 'study' : study_data,
         'trait' : trait_data, 
-        'variant' : get_variant_data, 
+        'variant' : variant_data, 
         # 'gene' : get_gene_data
     }
